Style: 이력서 템플릿 목록 UI 수정
